🚚 Rename: 게시물 상세 정보 가져오기 api 생성
