+++ src/app/(board)/board/[boardId]/_components/comment-list/comment.tsx
@@ -1,5 +1,5 @@
 import { IconProfile } from "@/public/assets/icons";
-import { ArticleCommentResponse } from "@/types/article/comment";
+import { ArticleCommentResponse } from "@/types/board/comment";
 import formatBoardDate from "@/utils/format-board-date";
 
 import CommentDropDown from "./comment-drop-down";
