+++ src/types/board/index.ts
@@ -1,13 +1,13 @@
-export type ArticleWriter = {
+export type BoardWriter = {
   nickname: string;
   id: number;
 };
 
-export type ArticleResponse = {
+export type BoardResponse = {
   updatedAt: string | Date;
   createdAt: string | Date;
   likeCount: number;
-  writer: ArticleWriter;
+  writer: BoardWriter;
   image: string | null;
   title: string;
   id: number;
